[TH-1209] 64비트 리뷰 ID 처리 수정

+++ app/src/main/java/com/zion830/threedollars/datasource/model/v2/response/my/Review.kt
@@ -11,7 +11,7 @@ data class Review(
     @SerializedName("rating")
     val rating: Float = 0f,
     @SerializedName("reviewId")
-    val reviewId: Int = 0,
+    val reviewId: Long = 0L,
     @SerializedName("updatedAt")
     val updatedAt: String = "",
     @SerializedName("user")
